fix(material/form-field): safely coerce line-height to em (#23215)
Our typography mixins for form-field related components coerce line-height to `em` units in several places. If a user provides a `line-height` via typography config with a unit specified, this results in an error as the result with be em squared. This fix changes the coercion to only occur if the line-height is unitless.