test(material/expansion): fix failing unit test on iOS14 safari
An expansion test seems to be flaky/failing in Safari on iOS14
because the `visibility: hidden` style from the expansion panel
content is not applied immediately. We fix this by enforcing
a style recalulcation like we do in the ripple. An alternative
would have been to switch the test to actual async/await where
we then need to wait at least until the next tick. It seems more
idiomatic to stick with `fakeAsync` here though.