fix(material/core): sanity checks not running (#23289)
Looks like when the granular sanity checks were introduced, some of the logic that determines whether a check should run wasn't written correctly which means that it isn't running at all.

These changes move the check logic into a separate method in order to make it less prone to mistakes.