build: group commits by package name in changelog (#23460)
Currently the new release tooling groups changes in the
changelog based on their commit scope. This is the default
behavior that works well for FW and the CLI repo.

For the components repository, we also include secondary entry-point
information in commit messages. This currently results in
many unnecessary groups, while ideally we'd want to group
based on the package name in the scope. This commit implements
that.