fix(material-experimental/mdc-typography): better mappings for 2014 to 2018 typography (#23618)
* fix(material-experimental/mdc-menu): fix bold font when using 2014 typography

* fix(material-experimental/mdc-typography): better mappings for 2014 to 2018 typography

This PR cleans up some typography hacks in menu and table in favor of a
higher level fix in the mapping code itself. Despite the naming
subheading-1 in the 2014 system looks more like body-1 in the 2018 system
and body-2 in the 2014 system looks more like subtitle-2 in the 2018
system. See my hackmd for the original typography mapping plan for a
side-by-side comparison of the different levels & systems:
https://hackmd.io/-IBynZC1QHCBefcFQyaAkA