build: generate linked ESM bundles of framework packages
As of APF v13, Angular packages come with partial compilation output.
This means that for tests, or running the dev-app we want to run the
linker to avoid JIT compilation of the FW code.

This commit introduces a macro for generating linked ESM bundles of
framework packages and their entry-points. These bundles can then
be passed later to tests or the dev-app, so that it would not need
to run the linker at all. This helps speeding up the development
turnaround as the linker would not need to re-run *always*..