build: remove init test environment calls from harness examples (#23853)
Removes the init test environment calls from the harness examples. This
is meant to make the examples a little more readable, and easier to
integrate within CLI projects (especially when are starting to use
a plain CLI project as foundation for our StackBlitz examples).

The CLI sets up the test environment in a global test init file, so
we should make this assumption in our examples as well. This also helps
with reducing duplication in our code base.