fix(material/tabs): pagination sometimes incorrectly shown after zoom (#23759)
Currently if one of the end tabs is selected and the user zooms in, we may end up showing the pagination unnecessarily. The issue comes from the fact that there's a transition on the ink bar which can cause the parent overflow while it is being measured.

These changes resolve the issue by measuring a different element.

Fixes #23724.