fix(autocomplete): restore focus after emitting option selected event (#18707)
Currently we restore focus to the input and then we emit the change event, but we have a report that it may be making some use cases more difficult. From what I can tell, this shouldn't have much of an impact on existing users so these changes swap the order so that the focus event is last.

Fixes #18650.