fix(material/slider): Ticks updated wrongly if the max property 0 (#24218)
* fix(material/slider): Ticks updated wrongly if the max property changed to 0

When we get equal min & max value for the slider then we can get division by zero which leads to an improper style value for the background-size (Infinity%). This fix introduces a tiny function that checks if we deal with finite numbers during calculations where potentially division by zero may occur.

Fixes #23913

* fix(material/slider): Check if interval persentage a valid number

Fixes #23913