fix(material/datepicker): content overflowing when large custom header is provided (#24255)
* fix(material/datepicker): content overflowing when large custom header is provided

Fixes a bug in the Angular Material `datepicker` component where the content of the datepicker is overflowing outside when a custom header with large height is provided. This is because of the fixed height provided is not enough with custom headers, so added a new class which will set the height of calendar to auto when custom header is given.

Fixes #20459

* fix(material/datepicker): content overflowing when large custom header is provided

    Fixes a bug in the Angular Material `datepicker` component where the content of the datepicker is overflowing outside when a custom header with large height is provided. This commit adds comments for the changes done for this bug.

    Fixes #20459

* fix(material/datepicker): content overflowing when large custom header is provided

Fixes a bug in the Angular Material `datepicker` component where the content of the datepicker is overflowing outside when a custom header with large height is provided. This commit updates the html to use plain class instead of ngClass and removes the $non-touch-custom-header-calendar-height scss variable.

Fixes #20459