fix(material/sidenav): prevent focus from entering hidden sidenav if child element has a visibility
This is a similar fix that we've made for other components. Usually we set `visibility: hidden` on the sidenav which is enough to prevent the user from tabbing into it, however one of the children could override it with `visibility: visible`. These changes add an extra `display: none` on top which will prevent such issues. This has come up on AIO recently https://github.com/angular/angular/pull/44990.